Code changes as per Thierry's comments
